rename pool_collector file
