+++ shared/src/lib.rs
@@ -6,7 +6,7 @@ pub mod current_block;
 pub mod gas_price_estimation;
 pub mod metrics;
 pub mod network;
-pub mod pool_collector;
+pub mod pool_aggregating;
 pub mod pool_fetching;
 pub mod price_estimate;
 pub mod time;
