+++ shared/src/pool_aggregating.rs
